Reorganize pkg/git: merge git, repository and extract the PR-specific parts.
