+++ pkg/git/git.go
@@ -1,92 +1,245 @@
 package git
 
 import (
-	"encoding/json"
+	"context"
 	"fmt"
-	"strings"
+	"log"
+	"path/filepath"
+	"time"
+
+	git2go "github.com/libgit2/git2go/v31"
 )
 
-const DefaultUsername = "git"
-const DefaultRemote = "origin"
-const DefaultBranch = "main"
-const PromoteBranchPrefix = "promote/"
+const (
+	DefaultUsername = "git"
+	DefaultRemote   = "origin"
+	DefaultBranch   = "main"
+)
 
-type Status struct {
+type CommitStatus struct {
 	Succeeded bool
 }
 
-type PullRequest struct {
-	ID          int
-	Title       string
-	Description string
-	State       PRState
+// Repository represents a local git repository.
+type Repository struct {
+	gitRepository *git2go.Repository
+	gitProvider   GitProvider
+	token         string
 }
 
-type PRState struct {
-	Env   string `json:"env"`
-	Group string `json:"group"`
-	App   string `json:"app"`
-	Tag   string `json:"tag"`
-	Sha   string `json:"sha"`
+// LoadRepository loads a local git repository.
+func LoadRepository(ctx context.Context, path string, providerTypeString string, token string) (*Repository, error) {
+	localRepo, err := git2go.OpenRepository(path)
+	if err != nil {
+		return &Repository{}, fmt.Errorf("could not open repository: %w", err)
+	}
+	remote, err := localRepo.Remotes.Lookup(DefaultRemote)
+	if err != nil {
+		return nil, fmt.Errorf("could not get remote: %w", err)
+	}
+	providerType, err := StringToProviderType(providerTypeString)
+	if err != nil {
+		return nil, fmt.Errorf("could not get provider type: %w", err)
+	}
+	provider, err := NewGitProvider(ctx, providerType, remote.Url(), token)
+	if err != nil {
+		return nil, fmt.Errorf("could not create git provider: %w", err)
+	}
+	return &Repository{
+		gitRepository: localRepo,
+		gitProvider:   provider,
+		token:         token,
+	}, nil
 }
 
-func (p *PRState) Title() string {
-	return fmt.Sprintf("Promote %s/%s in %s to %s", p.Group, p.App, p.Env, p.Tag)
+// GetRootDir returns the file path to the repository.
+func (g *Repository) GetRootDir() string {
+	p := g.gitRepository.Path()
+	rp := filepath.Clean(filepath.Join(p, ".."))
+	return rp
 }
 
-func (p *PRState) Description() (string, error) {
-	jsonString, err := json.Marshal(p)
+// CreateBranch creates a branch.
+func (g *Repository) CreateBranch(branchName string, force bool) error {
+	branch, err := g.gitRepository.LookupBranch(branchName, git2go.BranchLocal)
+	if err == nil {
+		err = branch.Delete()
+		if err != nil {
+			return fmt.Errorf("could not delete existing branch %q: %w", branchName, err)
+		}
+	}
+	head, err := g.gitRepository.Head()
+	if err != nil {
+		return err
+	}
+	headCommit, err := g.gitRepository.LookupCommit(head.Target())
+	if err != nil {
+		return err
+	}
+	_, err = g.gitRepository.CreateBranch(branchName, headCommit, force)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// GetLastCommitForBranch returns the latest commit id for the branch.
+func (g *Repository) GetLastCommitForBranch(branchName string) (*git2go.Oid, error) {
+	branch, err := g.gitRepository.LookupBranch(branchName, git2go.BranchAll)
+	if err != nil {
+		return nil, err
+	}
+	return branch.Target(), nil
+}
+
+// GetCurrentCommit returns the commmit id for HEAD.
+func (g *Repository) GetCurrentCommit() (*git2go.Oid, error) {
+	head, err := g.gitRepository.Head()
+	if err != nil {
+		return nil, err
+	}
+	return head.Target(), nil
+}
+
+// CreateCommit creates a commit in the specfied branch with the current changes.
+func (g *Repository) CreateCommit(branchName, message string) (*git2go.Oid, error) {
+	// TODO change to some bot name, probably break out in to config
+	signature := &git2go.Signature{
+		Name:  "gitops-promotion",
+		Email: "gitops-promotion@xenit.se",
+		When:  time.Now(),
+	}
+	idx, err := g.gitRepository.Index()
+	if err != nil {
+		return nil, err
+	}
+	err = idx.AddAll([]string{}, git2go.IndexAddDefault, nil)
+	if err != nil {
+		return nil, err
+	}
+	treeId, err := idx.WriteTree()
+	if err != nil {
+		return nil, err
+	}
+	err = idx.Write()
+	if err != nil {
+		return nil, err
+	}
+	tree, err := g.gitRepository.LookupTree(treeId)
+	if err != nil {
+		return nil, err
+	}
+	branch, err := g.gitRepository.LookupBranch(branchName, git2go.BranchLocal)
+	if err != nil {
+		return nil, err
+	}
+	commitTarget, err := g.gitRepository.LookupCommit(branch.Target())
+	if err != nil {
+		return nil, err
+	}
+	refName := fmt.Sprintf("refs/heads/%s", branchName)
+	sha, err := g.gitRepository.CreateCommit(refName, signature, signature, message, tree, commitTarget)
+	if err != nil {
+		return nil, err
+	}
+	log.Printf("Created commit %s on %s with message '%s'\n", sha, refName, message)
+	return sha, nil
+}
+
+// Push pushes the given branch to the remote.
+func (g *Repository) Push(branchName string, force bool) error {
+	remote, err := g.gitRepository.Remotes.Lookup(DefaultRemote)
+	if err != nil {
+		return fmt.Errorf("could not find remote %q: %w", DefaultRemote, err)
+	}
+
+	forceFlag := "+"
+	if !force {
+		forceFlag = ""
+	}
+
+	branches := []string{fmt.Sprintf("%srefs/heads/%s", forceFlag, branchName)}
+	err = remote.Push(branches, &git2go.PushOptions{RemoteCallbacks: credentialsCallback(DefaultUsername, g.token)})
+	if err != nil {
+		return fmt.Errorf("failed pushing branches %s: %w", branches, err)
+	}
+	log.Printf("Pushed branch %s to remote\n", branches[0])
+	return nil
+}
+
+// CreatePR creates a PR for the branch. It assumes that the branch has been pushed.
+func (g *Repository) CreatePR(ctx context.Context, branchName string, auto bool, state *PRState) (int, error) {
+	return g.gitProvider.CreatePR(ctx, branchName, auto, state)
+}
+
+// GetStatus returns the status for the give commit.
+func (g *Repository) GetStatus(ctx context.Context, sha, group, env string) (CommitStatus, error) {
+	return g.gitProvider.GetStatus(ctx, sha, group, env)
+}
+
+// MergePR merges the given PR.
+func (g *Repository) MergePR(ctx context.Context, id int, sha string) error {
+	return g.gitProvider.MergePR(ctx, id, sha)
+}
+
+// GetBranchName returns the branch name of for HEAD.
+func (g *Repository) GetBranchName() (string, error) {
+	head, err := g.gitRepository.Head()
 	if err != nil {
 		return "", err
 	}
-	description := fmt.Sprintf(`<!-- metadata = %s -->
-	ENV: %s
-	APP: %s
-	TAG: %s`, string(jsonString), p.Env, p.App, p.Tag)
-	return description, nil
-}
-
-func parsePrState(body string) (PRState, error) {
-	comp := strings.Split(body, " -->")
-	if len(comp) < 2 {
-		return PRState{}, fmt.Errorf("invalid metadata: %q", body)
-	}
-	comp = strings.Split(comp[0], "<!-- metadata = ")
-	if len(comp) < 2 {
-		return PRState{}, fmt.Errorf("invalid metadata: %q", body)
-	}
-	out := comp[1]
-	prState := PRState{}
-	err := json.Unmarshal([]byte(out), &prState)
+	branchName, err := head.Branch().Name()
 	if err != nil {
-		return PRState{}, err
+		return "", err
 	}
-	return prState, nil
+	return branchName, nil
 }
 
-func newPR(id *int, title *string, description *string, state *PRState) (PullRequest, error) {
-	if id == nil {
-		return PullRequest{}, fmt.Errorf("id can't be empty")
+// GetPRForCurrentBranch returns any active PR for the current branch.
+func (g *Repository) GetPRForCurrentBranch(ctx context.Context) (PullRequest, error) {
+	branchName, err := g.GetBranchName()
+	if err != nil {
+		return PullRequest{}, err
 	}
-
-	if title == nil {
-		return PullRequest{}, fmt.Errorf("title can't be empty")
+	pr, err := g.gitProvider.GetPRWithBranch(ctx, branchName, DefaultBranch)
+	if err != nil {
+		return PullRequest{}, err
 	}
+	return pr, nil
+}
 
-	d := ""
-	if description != nil {
-		d = *description
+// GetPRThatCausedCurrentCommit finds the merged PR with resulted in the current commit.
+func (g *Repository) GetPRThatCausedCurrentCommit(ctx context.Context) (PullRequest, error) {
+	head, err := g.gitRepository.Head()
+	if err != nil {
+		return PullRequest{}, err
 	}
-
-	s := PRState{}
-	if state != nil {
-		s = *state
+	pr, err := g.gitProvider.GetPRThatCausedCommit(ctx, head.Target().String())
+	if err != nil {
+		return PullRequest{}, err
 	}
+	return pr, err
+}
 
-	return PullRequest{
-		ID:          *id,
-		Title:       *title,
-		Description: d,
-		State:       s,
-	}, nil
+func Clone(url, username, password, path, branchName string) error {
+	_, err := git2go.Clone(url, path, &git2go.CloneOptions{
+		FetchOptions: &git2go.FetchOptions{
+			DownloadTags:    git2go.DownloadTagsNone,
+			RemoteCallbacks: credentialsCallback(username, password),
+		},
+		CheckoutBranch: branchName,
+	})
+	return err
+}
+
+func credentialsCallback(username, password string) git2go.RemoteCallbacks {
+	return git2go.RemoteCallbacks{
+		CredentialsCallback: func(url string, usernameFromURL string, allowedTypes git2go.CredentialType) (*git2go.Credential, error) {
+			cred, err := git2go.NewCredentialUserpassPlaintext(username, password)
+			if err != nil {
+				return nil, err
+			}
+			return cred, nil
+		},
+	}
 }
